% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGQ.R
\name{MGQ}
\alias{MGQ}
\title{MGQ}
\usage{
MGQ(
  num_items = NULL,
  type = "metal",
  with_welcome = TRUE,
  with_finish = FALSE,
  with_feedback = FALSE,
  label = "MGQ",
  dict = mpiquiz::mpiquiz_dict,
  timeout = 180,
  ...
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test. Default NULL pulls all items.}

\item{type}{(Character scalar) Type of quiz to implement. Currently supported: classical, metal, jazz, hiphop}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_finish}{(Logical scalar) Whether to show a finished page.}

\item{with_feedback}{(Logical scalar) Whether to include feedback to the participants.}

\item{label}{(Character scalar) Label to give the MGQ results in the output file.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{timeout}{(Double scalar) The time to answer (in seconds)}

\item{...}{Further arguments to be passed to \code{MGQ_main_test()}.}
}
\description{
This function defines a MGQ  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSM in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the MSM, consider using \code{\link{MGQ_demo}()}.
For a standalone implementation of the MGQ,
consider using \code{\link{MGQ_standalone}()}.
}
