% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGQ.R
\name{MGQ_demo}
\alias{MGQ_demo}
\title{Demo MGQ}
\usage{
MGQ_demo(
  num_items = 3L,
  type = "metal",
  timeout = 180,
  title = "MGQ Demo",
  dict = mpiquiz::mpiquiz_dict,
  admin_password = "demo",
  researcher_email = "klaus.frieler@ae.mpg.de",
  language = c("en", "de", "de_f")
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test. Default NULL pulls all items.}

\item{type}{(Character scalar) Type of quiz to implement. Currently supported: classical, metal, jazz, hiphop}

\item{timeout}{(Double scalar) The time to answer (in seconds)}

\item{title}{(Character scalar) The title}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{longgold@gold.uc.ak},
the email address of this package's developer.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"en"}) and German (informal: \code{"de"}, formal: \code{"de_f"}).
The first language is selected by default}

\item{...}{Further arguments to be passed to \code{\link{MGQ}()}.}
}
\description{
This function launches a demo for the MGQ
}
