% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGQ.R
\name{MGQ_standalone}
\alias{MGQ_standalone}
\title{Standalone MGQ}
\usage{
MGQ_standalone(
  title = NULL,
  num_items = NULL,
  type = "metal",
  timeout = 180,
  with_id = FALSE,
  with_welcome = TRUE,
  with_feedback = TRUE,
  admin_password = "conifer",
  researcher_email = "klaus.frieler@ae.mpg.de",
  languages = c("en", "de", "de_f"),
  dict = mpiquiz::mpiquiz_dict,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items to be adminstered. Default NULL pulls all items.}

\item{timeout}{(Double scalar) The time to answer (in seconds)}

\item{with_id}{(Logical scalar) Whether to show a ID page.}

\item{with_welcome}{(Logical scalar) Whether to show a welcome page.}

\item{with_feedback}{(Logical scalar) Whether to include feedback to the participants.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}),
and German (informal: \code{"DE"}, formal: \code{"DE_f"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{MGQ}()}.}
}
\description{
This function launches a standalone testing session for the MGQ
This can be used for data collection, either in the laboratory or online.
}
