% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_result.R
\name{parse_mgc_result}
\alias{parse_mgc_result}
\title{parse_mgc_result}
\usage{
parse_mgc_result(entry)
}
\arguments{
\item{entry}{(list) One entry of a psychTestR battery result list}
}
\description{
This function parses a entry in a typical psychTestR battery result list and returns it as a neat data frame
Will check for all quizzes at once.
}
